Refactor to follow Play terminology
